requirements.txt
appdirs==1.4.4
black==20.8b1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
coloredlogs==14.0
dataclasses==0.8
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
gspread==3.2.0
httplib2==0.17.0
humanfriendly==8.1
idna==2.9
mypy-extensions==0.4.3
numpy==1.18.1
oauth2client==4.1.3
oauthlib==3.1.0
pandas==1.0.1
pathspec==0.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2019.3
regex==2020.11.13
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.8
